Revert "ata: make transfer mode masks *unsigned int*"
This reverts commit f0a6d77b351c18c122fc1638ac9e58f5e0346f64.

+++ drivers/ata/pata_serverworks.c
@@ -150,7 +150,7 @@ static u8 serverworks_is_csb(struct pci_dev *pdev)
  *	bug we hit.
  */
 
-static unsigned int serverworks_osb4_filter(struct ata_device *adev, unsigned int mask)
+static unsigned long serverworks_osb4_filter(struct ata_device *adev, unsigned long mask)
 {
 	if (adev->class == ATA_DEV_ATA)
 		mask &= ~ATA_MASK_UDMA;
@@ -166,7 +166,7 @@ static unsigned int serverworks_osb4_filter(struct ata_device *adev, unsigned in
  *	Check the blacklist and disable UDMA5 if matched
  */
 
-static unsigned int serverworks_csb_filter(struct ata_device *adev, unsigned int mask)
+static unsigned long serverworks_csb_filter(struct ata_device *adev, unsigned long mask)
 {
 	const char *p;
 	char model_num[ATA_ID_PROD_LEN + 1];
